Needs testing: added squaring and resizing of images to preprocess step and removed grayscale and random crop
