+++ scripts/segmentation.py
@@ -13,115 +13,6 @@ from progressbar import ProgressBar
 
 from PIL import Image
 
-def optimal_square(x1,x2,y1,y2,i):
-	max_x = i.shape[1]
-	max_y = i.shape[0]
-
-	ret_x1 = x1
-	ret_x2 = x2
-	ret_y1 = y1
-	ret_y2 = y2
-
-	x_length = x2-x1
-	y_length = y2-y1
-	square_size = 0
-	affected_dimension = ''
-
-	side_1 = 0
-	side_2 = 0
-	delta = 0
-
-	if (x_length > y_length):
-		if x_length > max_y:
-			square_size = y_length
-			affected_dimension = 'x'
-			delta = abs(square_size - x_length)
-		else:
-			square_size = x_length
-			affected_dimension = 'y'
-			delta = abs(square_size - y_length)
-
-	elif (y_length > x_length):
-
-		if y_length > max_x:
-			square_size = x_length
-			affected_dimension = 'y'
-			delta = abs(square_size - y_length)
-		else:
-			square_size = y_length
-			affected_dimension = 'x'
-			delta = abs(square_size - x_length)
-
-	# print("Square Size: %d\n" % square_size)
-	# print("Affected_dimension: %s\n" % affected_dimension)
-	# print("Max Y: %d\n" % max_y)
-	# print("Max X: %d\n" % max_x)
-
-	# Split up addition/subtraction on both sides of proposed subimage
-	side_1 = int(round(delta / 2))
-	side_2 = delta - side_1
-
-	# print ("Side 1: %d\n" % side_1)
-	# print ("Side 2: %d\n" % side_2)
-
-	if affected_dimension == 'y':
-		if square_size > y_length:
-			if y1 < side_1:
-				side_2 = side_2 + (side_1-y1)
-
-				ret_y1 = 0
-				ret_y2 = y2 + side_2
-
-			elif (max_y - y2) > side_2:
-				side_1 = side_1 + (side_2 - (max_y-y2))
-
-				ret_y1 = y1 - side_1
-				ret_y2 = max_y
-
-			else:
-				ret_y1 = y1 - side_1
-				ret_y2 = y2 + side_2
-
-		elif square_size < y_length:
-			ret_y1 = y1 + side_1
-			ret_y2 = y2 - side_2
-
-	elif affected_dimension == 'x':
-		if square_size > x_length:
-			if x1 < side_1:
-				side_2 = side_2 + (side_1-x1)
-
-				ret_x1 = 0
-				ret_x2 = x2 + side_2
-
-			elif (max_x - x2) > side_2:
-				side_1 = side_1 + (side_2 - (max_x-x2))
-
-				ret_x1 = x1 - side_1
-				ret_x2 = max_x
-
-			else:
-				ret_x1 = x1 - side_1
-				ret_x2 = x2 + side_2
-
-		elif square_size < x_length:
-			ret_x1 = x1 + side_1
-			ret_x2 = x2 - side_2
-
-	# print ("Original X1 %d\n" % x1)
-	# print ("X1 %d\n" % ret_x1)
-	# print ("Original X2 %d\n" % x2)
-	# print ("X2 %d\n" % ret_x2)
-	# print ("Original Y1 %d\n" % y1)
-	# print ("Y1 %d\n" % ret_y1)
-	# print ("Original Y2 %d\n" % y2)
-	# print ("Y2 %d\n" % ret_y2)
-
-	# print("X length: %d\n" % (ret_x2-ret_x1))
-	# print("Y length: %d\n" % (ret_y2-ret_y1))
-
-	return int(ret_x1), int(ret_x2), int(ret_y1), int(ret_y2)
-
 def segment_images( burst_directory, output_directory, curr_datetime):
 	global latest_timestamp
 	pbar = ProgressBar()
@@ -196,7 +87,7 @@ def segment_images( burst_directory, output_directory, curr_datetime):
 			y1 = max(0, y1 - h*0.2)
 			y2 = min(i.shape[0]-1, y2 + h*0.8)
 
-			opt_x1, opt_x2, opt_y1, opt_y2 = optimal_square(int(x1),int(x2),int(y1),int(y2),i)
+			opt_x1, opt_x2, opt_y1, opt_y2 = tools.optimal_square(int(x1),int(x2),int(y1),int(y2),i)
 
 			expected_x = opt_x2 - opt_x1
 			expected_y = opt_y2 - opt_y1
